requirements.txt
numpy
Flask
gunicorn
scikit-learn
PyYAML
pandas

